// fubo-app/types/match.ts
export interface Match {
  id: string;
  title: string;
  hometeam: string;
  awayteam: string;
  hometeamabbr: string;
  awayteamabbr: string;
  hometeamID: string;
  awayteamID: string;
  starttime: string;
  endtime: string;
  sport: string;
  league: string;
  league_id: string;
  network: string;
  networkUrl: string;
  matchId: string;
  matchUrl: string;
  thumbnail: string;
  country: string;
  url: string;
  regionalRestrictions: boolean;
} 